separate store modules for user vs. tokenUser
