+++ static/js/src/entry/account/mixins/route.js
@@ -53,7 +53,7 @@ export default {
   },
 
   computed: {
-    ...mapState('user', ['accessToken']),
+    ...mapState('tokenUser', ['accessToken']),
   },
 
   methods: {
