+++ static/js/src/entry/account/mixins/token-user.js
@@ -2,6 +2,6 @@ import { mapState } from 'vuex';
 
 export default {
   computed: {
-    ...mapState('user', { tokenUser: 'tokenDetails' }),
+    ...mapState('tokenUser', { tokenUser: 'details' }),
   },
 };
