+++ static/js/src/entry/account/store/index.js
@@ -2,6 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 import userModule from './modules/user';
+import tokenUserModule from './modules/token-user';
 import contextModule from './modules/context';
 
 Vue.use(Vuex);
@@ -9,6 +10,7 @@ Vue.use(Vuex);
 const store = new Vuex.Store({
   modules: {
     user: userModule,
+    tokenUser: tokenUserModule,
     context: contextModule,
   },
 });
